Refactor to extract validation in three different classes
